+++ deltasigma/_circ_smooth.py
@@ -68,9 +68,9 @@ def circ_smooth(x, n=16):
         Amp = undbv(-3)
         f = 0.3
         N = 2**12
-        f1_bin = np.round(f1*N)
-        f2_bin = np.round(f2*N)
-        fin = np.round(((1 - f)/2*f1 + (f + 1)/2*f2) * N)
+        f1_bin = int(np.round(f1*N))
+        f2_bin = int(np.round(f2*N))
+        fin = int(np.round(((1 - f)/2*f1 + (f + 1)/2*f2) * N))
         t = np.arange(0, N)
         u = Amp*np.cos((2*np.pi/N)*fin*t)
         v, xn, xmax, y = simulateDSM(u, ntf, 2)
@@ -81,9 +81,9 @@ def circ_smooth(x, n=16):
         # plotting
         plt.figure(figsize=(12, 7))
         plt.subplot(211)
-        plt.plot(freq, dbv(spec0[:N/2 + 1]), 'c', linewidth=1, label='$S$')
+        plt.plot(freq, dbv(spec0[:int(N/2) + 1]), 'c', linewidth=1, label='$S$')
         spec_smoothed = circ_smooth(np.abs(spec0)**2., 16)
-        plt.plot(freq, dbp(spec_smoothed[:N/2 + 1]), 'b--', linewidth=2, label='$\\\\mathrm{circ\\\\_smooth}(S)$')
+        plt.plot(freq, dbp(spec_smoothed[:int(N/2) + 1]), 'b--', linewidth=2, label='$\\\\mathrm{circ\\\\_smooth}(S)$')
         plotSpectrum(spec0, fin, 'r', linewidth=2, label='$\\\\mathrm{plotSpectrum}(S)$')
         Snn = np.abs(evalTF(ntf, np.exp(2j*np.pi*freq)))**2 * 2/12*(delta)**2
         plt.plot(freq, dbp(Snn*NBW), 'm', linewidth=1.5, label='$\\\\mathrm{from\\\\ NTF}$')
@@ -100,9 +100,9 @@ def circ_smooth(x, n=16):
         ax.set_title('Smoothing and plotting for LOG and LIN axes')
         plt.legend(loc=4)
         plt.subplot(212)
-        plt.plot(freq, dbv(spec0[:N/2 + 1]), 'c', linewidth=1, label='$S$')
+        plt.plot(freq, dbv(spec0[:int(N/2) + 1]), 'c', linewidth=1, label='$S$')
         plotSpectrum(spec0, fin, '--r', linewidth=2, label='$\\\\mathrm{plotSpectrum}(S)$')
-        plt.plot(freq, dbp(spec_smoothed[:N/2 + 1]), 'b', linewidth=2, label='$\\\\mathrm{circ\\\\_smooth}(S)$')
+        plt.plot(freq, dbp(spec_smoothed[:int(N/2) + 1]), 'b', linewidth=2, label='$\\\\mathrm{circ\\\\_smooth}(S)$')
         plt.plot(freq, dbp(Snn*NBW), 'm', linewidth=1.5, label='$\\\\mathrm{from\\\\ NTF}$')
         msg = 'SQNR  =  %.1fdB\\n @ A = %.1fdBFS & osr = %.0f\\n' % \\
               (snr, dbv(spec0[fin]), osr)
