+++ deltasigma/_logsmooth.py
@@ -85,9 +85,9 @@ def logsmooth(X, inBin, nbin=8, n=3):
         u0 = u0 + .01*u0**2+.001*u0**3+.005*u0**4
         U = np.fft.fft(u0 * ds_hann(N))/(N/4)
         f = np.linspace(0, Fs, N + 1)
-        f = f[:N/2 + 1]
+        f = f[:int(N/2) + 1]
         plt.subplot(211)
-        plt.semilogx(f, dbv(U[:N/2 + 1]))
+        plt.semilogx(f, dbv(U[:int(N/2) + 1]))
         inBin = np.round(freq/Fs*N)
         fS, US = logsmooth(U, inBin)
         plt.semilogx(fS*Fs, US, 'r', linewidth=2.5)
